🚚(exceptions) move invitation-related exception to core
move invitation-related from mailbox_manager to core, as
it will be useful for teams too
