+++ src/backend/core/exceptions.py
